[cs] throw on FileSystem.deleteDirectory(path) and deleteFile(path) if path does not exist

+++ std/cs/_std/sys/FileSystem.hx
@@ -103,11 +103,13 @@ class FileSystem {
 
 	public static function deleteFile( path : String ) : Void
 	{
+		if (!File.Exists(path)) throw "Path '" + path + "' doesn't exist";
 		File.Delete(path);
 	}
 
 	public static function deleteDirectory( path : String ) : Void
 	{
+		if (!File.Exists(path)) throw "Path '" + path + "' doesn't exist";
 		Directory.Delete(path);
 	}
